refactor(portal): drain team-leaf GraphQL docs to scenes/common
