--- web/scenes/Portal/Teams/TeamId/Team/page/Apps/graphql/client/fetch-apps.generated.ts
@@ -1,112 +1,0 @@
-/* eslint-disable */
-import * as Types from "@/graphql/graphql";
-
-import { gql } from "@apollo/client";
-import * as Apollo from "@apollo/client";
-const defaultOptions = {} as const;
-export type FetchAppsQueryVariables = Types.Exact<{
-  teamId: Types.Scalars["String"]["input"];
-}>;
-
-export type FetchAppsQuery = {
-  __typename?: "query_root";
-  app: Array<{
-    __typename?: "app";
-    id: string;
-    is_staging: boolean;
-    engine: string;
-    app_metadata: Array<{
-      __typename?: "app_metadata";
-      id: string;
-      name: string;
-      logo_img_url: string;
-      verification_status: string;
-    }>;
-  }>;
-};
-
-export const FetchAppsDocument = gql`
-  query FetchApps($teamId: String!) {
-    app(where: { team_id: { _eq: $teamId } }) {
-      id
-      is_staging
-      engine
-      app_metadata {
-        id
-        name
-        logo_img_url
-        verification_status
-      }
-    }
-  }
-`;
-
-/**
- * __useFetchAppsQuery__
- *
- * To run a query within a React component, call `useFetchAppsQuery` and pass it any options that fit your needs.
- * When your component renders, `useFetchAppsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useFetchAppsQuery({
- *   variables: {
- *      teamId: // value for 'teamId'
- *   },
- * });
- */
-export function useFetchAppsQuery(
-  baseOptions: Apollo.QueryHookOptions<
-    FetchAppsQuery,
-    FetchAppsQueryVariables
-  > &
-    (
-      | { variables: FetchAppsQueryVariables; skip?: boolean }
-      | { skip: boolean }
-    ),
-) {
-  const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useQuery<FetchAppsQuery, FetchAppsQueryVariables>(
-    FetchAppsDocument,
-    options,
-  );
-}
-export function useFetchAppsLazyQuery(
-  baseOptions?: Apollo.LazyQueryHookOptions<
-    FetchAppsQuery,
-    FetchAppsQueryVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useLazyQuery<FetchAppsQuery, FetchAppsQueryVariables>(
-    FetchAppsDocument,
-    options,
-  );
-}
-export function useFetchAppsSuspenseQuery(
-  baseOptions?:
-    | Apollo.SkipToken
-    | Apollo.SuspenseQueryHookOptions<FetchAppsQuery, FetchAppsQueryVariables>,
-) {
-  const options =
-    baseOptions === Apollo.skipToken
-      ? baseOptions
-      : { ...defaultOptions, ...baseOptions };
-  return Apollo.useSuspenseQuery<FetchAppsQuery, FetchAppsQueryVariables>(
-    FetchAppsDocument,
-    options,
-  );
-}
-export type FetchAppsQueryHookResult = ReturnType<typeof useFetchAppsQuery>;
-export type FetchAppsLazyQueryHookResult = ReturnType<
-  typeof useFetchAppsLazyQuery
->;
-export type FetchAppsSuspenseQueryHookResult = ReturnType<
-  typeof useFetchAppsSuspenseQuery
->;
-export type FetchAppsQueryResult = Apollo.QueryResult<
-  FetchAppsQuery,
-  FetchAppsQueryVariables
->;
