+++ web/scenes/Portal/Teams/TeamId/Team/page/Members/List/RemoveUserDialog/index.tsx
@@ -12,8 +12,8 @@ import { useParams } from "next/navigation";
 import { useCallback } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "react-toastify";
-import { FetchTeamMembersDocument } from "../../graphql/client/fetch-team-members.generated";
-import { useRemoveUserMutation } from "./graphql/client/remove-user.generated";
+import { FetchTeamMembersDocument } from "@/scenes/common/team-members/graphql/client/fetch-team-members.generated";
+import { useRemoveUserMutation } from "@/scenes/common/team-members/graphql/client/remove-user.generated";
 
 export const removeUserDialogAtom = atom(false);
