+++ web/scenes/Portal/Teams/TeamId/Team/page/Members/List/index.tsx
@@ -9,16 +9,16 @@ import { useAtom } from "jotai";
 import { useParams } from "next/navigation";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { toast } from "react-toastify";
-import { useInviteTeamMembersMutation } from "../graphql/client/invite-team-members.generated";
+import { useInviteTeamMembersMutation } from "@/scenes/common/team-members/graphql/client/invite-team-members.generated";
 import { EditRoleDialog, editRoleDialogAtom } from "./EditRoleDialog";
 import { PermissionsDialog } from "./PermissionsDialog";
 import { RemoveUserDialog, removeUserDialogAtom } from "./RemoveUserDialog";
-import { useDeleteInviteMutation } from "./graphql/client/delete-invite.generated";
+import { useDeleteInviteMutation } from "@/scenes/common/team-members/graphql/client/delete-invite.generated";
 import {
   FetchTeamMembersDocument,
   FetchTeamMembersQuery,
   useFetchTeamMembersQuery,
-} from "../graphql/client/fetch-team-members.generated";
+} from "@/scenes/common/team-members/graphql/client/fetch-team-members.generated";
 import { Item } from "./Item";
 
 type ListProps = {
