+++ web/scenes/common/team-members/graphql/client/delete-invite.graphql
