+++ web/scenes/common/team-members/graphql/client/remove-user.graphql
